requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

requests
azure-functions
azure-functions-durable
langchain-community
pymupdf
azure-storage-blob
python-dotenv
langchain-openai
azure-identity
azure-search-documents